requirement.txt
blinker==1.9.0
celery==5.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ffmpeg-python==0.2.0
filelock==3.24.3
Flask==3.1.3
fsspec==2026.2.0
future==1.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai-whisper==20250625
redis==5.2.1
regex==2026.2.19
requests==2.32.5
setuptools==82.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
